moved all content from asciimath directory to math directory
