--- lib/plurimath/asciimath/function.rb
@@ -1,3 +1,0 @@
-require_relative "../math/function/sin"
-require_relative "../math/function/tan"
-require_relative "../math/function/cos"
