+++ lib/plurimath/math/function.rb
@@ -1,0 +1,3 @@
+require_relative "function/sin"
+require_relative "function/tan"
+require_relative "function/cos"
